backend/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Alexey Bobrov <bobrov.bobrov-alex@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.6"
pillow = "^9.4.0"
python-decouple = "^3.7"
psycopg2-binary = "^2.9.5"
dj-database-url = "^1.2.0"
gunicorn = "^20.1.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
drf-spectacular = "^0.25.1"
drf-spectacular-sidecar = "^2022.12.1"
channels = {extras = ["daphne"], version = "^4.0.0"}
channels-redis = "^4.0.0"
requests = "^2.28.2"
uvicorn = "^0.20.0"
aiohttp = "^3.8.4"
daphne = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
